Refactor triedb utils
