+++ monad-triedb-utils/examples/async.rs
@@ -4,7 +4,7 @@ use alloy_primitives::B256;
 use clap::Parser;
 use monad_eth_testutil::secret_to_eth_address;
 use monad_eth_types::EthAddress;
-use monad_triedb::Handle;
+use monad_triedb_utils::TriedbReader;
 
 #[derive(Debug, Parser)]
 struct Args {
@@ -17,16 +17,16 @@ fn main() -> Result<(), Box<dyn Error>> {
 
     let triedb_path = args.triedb_path.unwrap_or(PathBuf::new());
 
-    let handle = Handle::try_new(triedb_path.as_path()).unwrap();
+    let reader = TriedbReader::try_new(triedb_path.as_path()).unwrap();
 
-    let block_id = handle.latest_block();
+    let block_id = reader.get_latest_block();
     println!("latest block id: {}", block_id);
 
     let num_accounts = 1000;
     let eth_addresses: Vec<EthAddress> = (0..num_accounts)
         .map(|_| secret_to_eth_address(B256::random()))
         .collect();
-    let results = handle
+    let results = reader
         .get_accounts_async(eth_addresses.iter(), block_id)
         .unwrap();
     println!("{:?}", results);
