+++ monad-triedb-utils/src/lib.rs
@@ -1,350 +1,173 @@
 use std::{
-    cell::RefCell,
-    path::{Path, PathBuf},
+    path::Path,
+    sync::{
+        atomic::{AtomicUsize, Ordering::SeqCst},
+        Arc,
+    },
 };
 
-use alloy_primitives::{keccak256, U256};
-use alloy_rlp::{Decodable, Encodable};
-use monad_blockdb::BlockTagKey;
-use monad_blockdb_utils::BlockTags;
-use monad_triedb::{key::create_addr_key, Handle};
-use tracing::debug;
+use decode::rlp_decode_storage_slot;
+use futures::{executor::block_on, future::join_all, FutureExt};
+use monad_eth_types::{EthAccount, EthAddress};
+use monad_state_backend::{StateBackend, StateBackendError};
+use monad_triedb::Handle;
+use monad_types::SeqNum;
+use tracing::{debug, warn};
 
-pub type EthAddress = [u8; 20];
-pub type EthStorageKey = [u8; 32];
+use crate::{
+    decode::rlp_decode_account,
+    key::{create_addr_key, create_code_key, create_receipt_key, create_storage_at_key},
+};
+
+pub mod decode;
+pub mod key;
+
+const MAX_TRIEDB_ASYNC_POLLS: u32 = 320_000;
 
 #[derive(Clone)]
-pub struct TriedbEnv {
-    triedb_path: PathBuf,
+pub struct TriedbReader {
+    handle: Handle,
 }
 
-#[derive(Debug)]
-pub enum TriedbResult {
-    Null,
-    EncodingError,
-    // (nonce, balance, code_hash)
-    Account(u128, u128, [u8; 32]),
-    Storage([u8; 32]),
-    Code(Vec<u8>),
-    Receipt(Vec<u8>),
-    BlockNum(u64),
-}
-
-impl TriedbEnv {
-    thread_local! {
-        pub static DB: RefCell<Option<Handle>> = RefCell::new(None);
+impl TriedbReader {
+    pub fn try_new(triedb_path: &Path) -> Option<Self> {
+        Handle::try_new(triedb_path).map(|handle| Self { handle })
     }
 
-    pub fn new(triedb_path: &Path) -> Self {
-        Self {
-            triedb_path: triedb_path.to_path_buf(),
-        }
+    pub fn get_latest_block(&self) -> u64 {
+        self.handle.latest_block()
     }
 
-    pub async fn get_latest_block(&self) -> TriedbResult {
-        let triedb_path = self.triedb_path.clone();
-        let (send, recv) = tokio::sync::oneshot::channel();
-        rayon::spawn(move || {
-            TriedbEnv::DB.with_borrow_mut(|db_handle| {
-                let db = db_handle.get_or_insert_with(|| {
-                    TriedbEnv::new_conn(&triedb_path).expect("triedb should exist in path")
-                });
+    pub fn get_account(&self, eth_address: &[u8; 20], block_id: u64) -> Option<EthAccount> {
+        let (triedb_key, key_len_nibbles) = create_addr_key(eth_address);
 
-                let result = TriedbEnv::latest_block(db);
-                let _ = send.send(TriedbResult::BlockNum(result));
-            });
-        });
-        recv.await.expect("rayon panic get_latest_block")
+        let result = self.handle.read(&triedb_key, key_len_nibbles, block_id);
+
+        let Some(account_rlp) = result else {
+            debug!("account {:?} not found at {:?}", eth_address, block_id);
+            return None;
+        };
+
+        rlp_decode_account(account_rlp)
     }
 
-    pub async fn get_account(&self, addr: EthAddress, block_tag: BlockTags) -> TriedbResult {
-        let triedb_path = self.triedb_path.clone();
-        let (send, recv) = tokio::sync::oneshot::channel();
-        rayon::spawn(move || {
-            TriedbEnv::DB.with_borrow_mut(|db_handle| {
-                let db = db_handle.get_or_insert_with(|| {
-                    TriedbEnv::new_conn(&triedb_path).expect("triedb should exist in path")
-                });
-                let (triedb_key, key_len_nibbles) = create_addr_key(&addr);
-
-                // parse block tag
-                let block_num = match block_tag {
-                    BlockTags::Number(q) => q,
-                    BlockTags::Default(t) => match t {
-                        BlockTagKey::Latest => TriedbEnv::latest_block(db),
-                        BlockTagKey::Finalized => TriedbEnv::latest_block(db),
-                    },
-                };
-
-                let result = TriedbEnv::read(db, &triedb_key, key_len_nibbles, block_num);
-                let Some(result) = result else {
-                    let _ = send.send(TriedbResult::Null);
-                    return;
-                };
-
-                let mut buf = result.as_slice();
-                let Ok(mut buf) = alloy_rlp::Header::decode_bytes(&mut buf, true) else {
-                    debug!("rlp decode failed: {:?}", buf);
-                    let _ = send.send(TriedbResult::EncodingError);
-                    return;
-                };
-
-                // address (currently not needed)
-                let Ok(_) = EthAddress::decode(&mut buf) else {
-                    debug!("rlp address decode failed: {:?}", buf);
-                    let _ = send.send(TriedbResult::EncodingError);
-                    return;
-                };
-
-                // account incarnation decode (currently not needed)
-                let Ok(_) = u64::decode(&mut buf) else {
-                    debug!("rlp incarnation decode failed: {:?}", buf);
-                    let _ = send.send(TriedbResult::EncodingError);
-                    return;
-                };
-
-                let Ok(nonce) = u128::decode(&mut buf) else {
-                    debug!("rlp nonce decode failed: {:?}", buf);
-                    let _ = send.send(TriedbResult::EncodingError);
-                    return;
-                };
-
-                let Ok(balance) = u128::decode(&mut buf) else {
-                    debug!("rlp balance decode failed: {:?}", buf);
-                    let _ = send.send(TriedbResult::EncodingError);
-                    return;
-                };
-
-                let code_hash = if buf.is_empty() {
-                    [0; 32]
-                } else {
-                    match <[u8; 32]>::decode(&mut buf) {
-                        Ok(x) => x,
-                        Err(e) => {
-                            debug!("rlp code_hash decode failed: {:?}", e);
-                            [0; 32]
-                        }
-                    }
-                };
-
-                let _ = send.send(TriedbResult::Account(nonce, balance, code_hash));
-            });
-        });
-        recv.await.expect("rayon panic get_account")
-    }
-
-    pub async fn get_storage_at(
+    pub fn get_accounts_async<'a>(
         &self,
-        addr: EthAddress,
-        at: EthStorageKey,
-        block_tag: BlockTags,
-    ) -> TriedbResult {
-        let triedb_path = self.triedb_path.clone();
-        let (send, recv) = tokio::sync::oneshot::channel();
-        rayon::spawn(move || {
-            TriedbEnv::DB.with_borrow_mut(|db_handle| {
-                let db = db_handle.get_or_insert_with(|| {
-                    TriedbEnv::new_conn(&triedb_path).expect("triedb should exist in path")
-                });
-                let (triedb_key, key_len_nibbles) = TriedbEnv::create_storage_at_key(&addr, &at);
+        eth_addresses: impl Iterator<Item = &'a EthAddress>,
+        block_id: u64,
+    ) -> Option<Vec<Option<EthAccount>>> {
+        // Counter which is updated when TrieDB processes a single async read to completion
+        let completed_counter = Arc::new(AtomicUsize::new(0));
 
-                // parse block tag
-                let block_num = match block_tag {
-                    BlockTags::Number(q) => q,
-                    BlockTags::Default(t) => match t {
-                        BlockTagKey::Latest => TriedbEnv::latest_block(db),
-                        BlockTagKey::Finalized => TriedbEnv::latest_block(db),
-                    },
-                };
-
-                let result = TriedbEnv::read(db, &triedb_key, key_len_nibbles, block_num);
-                let Some(result) = result else {
-                    let _ = send.send(TriedbResult::Null);
-                    return;
-                };
-
-                let mut buf = result.as_slice();
-                let Ok(mut buf) = alloy_rlp::Header::decode_bytes(&mut buf, true) else {
-                    debug!("rlp decode failed: {:?}", buf);
-                    let _ = send.send(TriedbResult::EncodingError);
-                    return;
-                };
-
-                // storage key (currently not needed)
-                let Ok(_) = U256::decode(&mut buf) else {
-                    debug!("rlp storage key decode failed: {:?}", buf);
-                    let _ = send.send(TriedbResult::EncodingError);
-                    return;
-                };
-
-                // storage value
-                let _ = match U256::decode(&mut buf) {
-                    Ok(res) => {
-                        let mut storage_value = [0_u8; 32];
-                        for (byte, storage) in res
-                            .to_be_bytes_vec()
-                            .into_iter()
-                            .zip(storage_value.iter_mut())
-                        {
-                            *storage = byte;
-                        }
-                        send.send(TriedbResult::Storage(storage_value))
-                    }
-                    Err(e) => {
-                        debug!("rlp storage value decode failed: {:?}", e);
-                        send.send(TriedbResult::EncodingError)
-                    }
-                };
-            });
+        let mut num_accounts = 0;
+        let eth_account_receivers = eth_addresses.map(|eth_address| {
+            num_accounts += 1;
+            let (triedb_key, key_len_nibbles) = create_addr_key(eth_address.as_ref());
+            self.handle
+                .read_async(
+                    triedb_key.as_ref(),
+                    key_len_nibbles,
+                    block_id,
+                    completed_counter.clone(),
+                )
+                .map(|receiver_result| {
+                    // Receiver should not fail
+                    let maybe_rlp_account = receiver_result.expect("receiver can't be canceled");
+                    // RLP decode the received account value
+                    maybe_rlp_account.and_then(rlp_decode_account)
+                })
         });
-        recv.await.expect("rayon panic get_storage_at")
-    }
 
-    pub async fn get_code(&self, code_hash: [u8; 32], block_tag: BlockTags) -> TriedbResult {
-        let triedb_path = self.triedb_path.clone();
-        let (send, recv) = tokio::sync::oneshot::channel();
-        rayon::spawn(move || {
-            TriedbEnv::DB.with_borrow_mut(|db_handle| {
-                let db = db_handle.get_or_insert_with(|| {
-                    TriedbEnv::new_conn(&triedb_path).expect("triedb should exist in path")
-                });
-                let (triedb_key, key_len_nibbles) = TriedbEnv::create_code_key(&code_hash);
+        // Join all futures of receivers
+        let eth_account_receivers = join_all(eth_account_receivers);
 
-                // parse block tag
-                let block_num = match block_tag {
-                    BlockTags::Number(q) => q,
-                    BlockTags::Default(t) => match t {
-                        BlockTagKey::Latest => TriedbEnv::latest_block(db),
-                        BlockTagKey::Finalized => TriedbEnv::latest_block(db),
-                    },
-                };
-
-                let result = TriedbEnv::read(db, &triedb_key, key_len_nibbles, block_num);
-                let Some(result) = result else {
-                    let _ = send.send(TriedbResult::Null);
-                    return;
-                };
-
-                let _ = send.send(TriedbResult::Code(result));
-            });
-        });
-        recv.await.expect("rayon panic get_code")
-    }
-
-    pub async fn get_receipt(&self, txn_index: u64, block_num: u64) -> TriedbResult {
-        let triedb_path = self.triedb_path.clone();
-        let (send, recv) = tokio::sync::oneshot::channel();
-        rayon::spawn(move || {
-            TriedbEnv::DB.with_borrow_mut(|db_handle| {
-                let db = db_handle.get_or_insert_with(|| {
-                    TriedbEnv::new_conn(&triedb_path).expect("triedb should exist in path")
-                });
-                let (triedb_key, key_len_nibbles) = TriedbEnv::create_receipt_key(txn_index);
-
-                let result = TriedbEnv::read(db, &triedb_key, key_len_nibbles, block_num);
-                let Some(result) = result else {
-                    let _ = send.send(TriedbResult::Null);
-                    return;
-                };
-
-                let _ = send.send(TriedbResult::Receipt(result));
-            });
-        });
-        recv.await.expect("rayon panic get_receipt")
-    }
-
-    fn new_conn(triedb_path: &Path) -> Option<Handle> {
-        Handle::try_new(triedb_path)
-    }
-
-    fn read(handle: &Handle, key: &[u8], key_len_nibbles: u8, block_id: u64) -> Option<Vec<u8>> {
-        handle.read(key, key_len_nibbles, block_id)
-    }
-
-    fn latest_block(handle: &Handle) -> u64 {
-        handle.latest_block()
-    }
-
-    fn create_storage_at_key(addr: &EthAddress, at: &EthStorageKey) -> (Vec<u8>, u8) {
-        let mut key_nibbles: Vec<u8> = vec![];
-
-        let state_nibble = 0_u8;
-        key_nibbles.push(state_nibble);
-
-        let hashed_addr = keccak256(addr);
-        for byte in &hashed_addr {
-            key_nibbles.push(*byte >> 4);
-            key_nibbles.push(*byte & 0xF);
+        let mut poll_count = 0;
+        // Poll TrieDB until the completed_counter reaches num_accounts
+        while completed_counter.load(SeqCst) < num_accounts && poll_count < MAX_TRIEDB_ASYNC_POLLS {
+            self.handle.triedb_poll();
+            poll_count += 1;
+        }
+        // TrieDB should have completed processing all the async callbacks at this point
+        if completed_counter.load(SeqCst) != num_accounts {
+            warn!(
+                "TrieDB poll count went over limit for {} lookups",
+                num_accounts
+            );
+            return None;
         }
 
-        let hashed_at = keccak256(at);
-        for byte in &hashed_at {
-            key_nibbles.push(*byte >> 4);
-            key_nibbles.push(*byte & 0xF);
-        }
-
-        let num_nibbles: u8 = key_nibbles.len().try_into().expect("key too big");
-        if num_nibbles % 2 != 0 {
-            key_nibbles.push(0);
-        }
-        let key: Vec<_> = key_nibbles
-            .chunks(2)
-            .map(|chunk| (chunk[0] << 4) | chunk[1])
-            .collect();
-
-        (key, num_nibbles)
+        Some(block_on(eth_account_receivers))
     }
 
-    fn create_code_key(code_hash: &[u8; 32]) -> (Vec<u8>, u8) {
-        let mut key_nibbles: Vec<u8> = vec![];
+    pub fn get_storage_at(
+        &self,
+        eth_address: &[u8; 20],
+        storage_key: &[u8; 32],
+        block_id: u64,
+    ) -> Option<[u8; 32]> {
+        let (triedb_key, key_len_nibbles) = create_storage_at_key(eth_address, storage_key);
 
-        let code_nibble = 1_u8;
-        key_nibbles.push(code_nibble);
+        let result = self.handle.read(&triedb_key, key_len_nibbles, block_id);
 
-        for byte in code_hash {
-            key_nibbles.push(*byte >> 4);
-            key_nibbles.push(*byte & 0xF);
-        }
+        let Some(storage_rlp) = result else {
+            debug!("storage {:?} not found at {:?}", eth_address, block_id);
+            return None;
+        };
 
-        let num_nibbles: u8 = key_nibbles.len().try_into().expect("key too big");
-        if num_nibbles % 2 != 0 {
-            key_nibbles.push(0);
-        }
-        let key: Vec<_> = key_nibbles
-            .chunks(2)
-            .map(|chunk| (chunk[0] << 4) | chunk[1])
-            .collect();
-
-        (key, num_nibbles)
+        rlp_decode_storage_slot(storage_rlp)
     }
 
-    pub fn path(&self) -> PathBuf {
-        self.triedb_path.clone()
+    pub fn get_code(&self, code_hash: &[u8; 32], block_id: u64) -> Option<Vec<u8>> {
+        let (triedb_key, key_len_nibbles) = create_code_key(code_hash);
+
+        self.handle.read(&triedb_key, key_len_nibbles, block_id)
     }
 
-    fn create_receipt_key(txn_index: u64) -> (Vec<u8>, u8) {
-        let mut key_nibbles: Vec<u8> = vec![];
+    pub fn get_receipt(&self, txn_index: u64, block_id: u64) -> Option<Vec<u8>> {
+        let (triedb_key, key_len_nibbles) = create_receipt_key(txn_index);
 
-        let receipt_nibble = 2_u8;
-        key_nibbles.push(receipt_nibble);
+        self.handle.read(&triedb_key, key_len_nibbles, block_id)
+    }
+}
 
-        let mut rlp_buf = vec![];
-        txn_index.encode(&mut rlp_buf);
-
-        for byte in rlp_buf {
-            key_nibbles.push(byte >> 4);
-            key_nibbles.push(byte & 0xF);
+impl StateBackend for TriedbReader {
+    fn get_account_statuses<'a>(
+        &self,
+        block: SeqNum,
+        eth_addresses: impl Iterator<Item = &'a EthAddress>,
+    ) -> Result<Vec<Option<EthAccount>>, StateBackendError> {
+        let latest = self.raw_read_latest_block();
+        if latest < block {
+            // latest < block
+            return Err(StateBackendError::NotAvailableYet);
         }
+        // block <= latest
 
-        let num_nibbles: u8 = key_nibbles.len().try_into().expect("key too big");
-        if num_nibbles % 2 != 0 {
-            key_nibbles.push(0);
+        let earliest = self.raw_read_earliest_block();
+        if block < earliest {
+            // block < earliest
+            return Err(StateBackendError::NeverAvailable);
         }
-        let key: Vec<_> = key_nibbles
-            .chunks(2)
-            .map(|chunk| (chunk[0] << 4) | chunk[1])
-            .collect();
+        // block >= earliest
 
-        (key, num_nibbles)
+        let Some(statuses) = self.get_accounts_async(eth_addresses, block.0) else {
+            // TODO: Use a more descriptive error
+            return Err(StateBackendError::NotAvailableYet);
+        };
+
+        // all accounts are now guaranteed to be fully consistent and correct
+        Ok(statuses)
+    }
+
+    fn raw_read_account(&self, block: SeqNum, address: &EthAddress) -> Option<EthAccount> {
+        self.get_account(address.as_ref(), block.0)
+    }
+
+    fn raw_read_earliest_block(&self) -> SeqNum {
+        SeqNum(self.handle.earliest_block())
+    }
+
+    fn raw_read_latest_block(&self) -> SeqNum {
+        SeqNum(self.handle.latest_block())
     }
 }
