+++ monad-triedb/Cargo.toml
@@ -9,19 +9,8 @@ bench = false
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-monad-state-backend = { path = "../monad-state-backend" }
-monad-eth-types = { path = "../monad-eth-types" }
-monad-types = { path = "../monad-types" }
-
-alloy-primitives = { workspace = true }
-alloy-rlp = { workspace = true }
 futures = { workspace = true }
 tracing = { workspace = true }
-
-[dev-dependencies]
-monad-eth-testutil = { path = "../monad-eth-testutil" }
-
-clap = { workspace = true, features = ["derive"] }
 
 [build-dependencies]
 bindgen = { workspace = true }
